flask_api/requirements.txt
flask>=2.3.2
flask-cors>=4.0.0
cassandra-driver>=3.25.0
python-dotenv>=1.0.0 